requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
Pillow==10.4.0
django-filter==24.1
python-decouple==3.8
